feat: Make producer-dashboard archive code visible (as lib)
